fix: feed mapper, email

+++ src/main/java/com/kosa/mapbegood/domain/mymap/myplaceFeed/service/MyplaceFeedService.java
@@ -1,35 +1,28 @@
 package com.kosa.mapbegood.domain.mymap.myplaceFeed.service;
 
-import com.kosa.mapbegood.domain.member.entity.Member;
 import com.kosa.mapbegood.domain.mymap.myplaceFeed.dto.MyplaceFeedDTO;
 import com.kosa.mapbegood.domain.mymap.myplaceFeed.entity.MyplaceFeed;
+import com.kosa.mapbegood.domain.mymap.myplaceFeed.mapper.MyplaceFeedMapper;
 import com.kosa.mapbegood.domain.mymap.myplaceFeed.repository.MyplaceFeedRepository;
 import com.kosa.mapbegood.exception.AddException;
 import com.kosa.mapbegood.exception.FindException;
 import com.kosa.mapbegood.exception.ModifyException;
 import com.kosa.mapbegood.exception.RemoveException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
 
 @Service
-public class MyplaceFeedService {
+public class MyplaceFeedService implements MyplaceFeedServiceInterface {
 
     @Autowired
     private MyplaceFeedRepository mfr;
 
-    /**
-     * DTO 객체 타입의 피드 작성자 닉네임을 entity 타입으로 변환한다
-     * @param feedDto 피드
-     * @return 멤버 엔티티
-     */
-    public Member getMemberNickname(MyplaceFeedDTO feedDto){
-        return Member
-                .builder()
-//                .nickname(feedDto.getMemberNickname())
-                .build();
-    }
+    @Autowired
+    private MyplaceFeedMapper mapper;
 
     /**
      * 내 피드를 조회할 수 있다
@@ -37,14 +30,13 @@ public class MyplaceFeedService {
      * @return 피드
      * @throws FindException
      */
-    public MyplaceFeedDTO findMyFeedById(Long myplaceId) throws FindException {
-        Optional<MyplaceFeed> feed = mfr.findById(myplaceId);
-        MyplaceFeedDTO feedDto = new MyplaceFeedDTO();
-        feedDto.setMyplaceId(myplaceId);
-//        feedDto.setMemberNickname(String.valueOf(feed.get().getMemberNickname()));
-        feedDto.setContent(feed.get().getContent());
-        feedDto.setCreatedAt(feed.get().getCreatedAt());
-        return feedDto;
+    public MyplaceFeed findMyFeedById(Long myplaceId) throws FindException {
+            Optional<MyplaceFeed> feed = mfr.findById(myplaceId);
+            if (feed.isPresent()) {
+                return feed.get();
+            } else {
+                throw new FindException("해당 피드를 찾을 수 없습니다");
+            }
     }
 
     /**
@@ -53,14 +45,12 @@ public class MyplaceFeedService {
      * @throws AddException
      */
     public void createMyFeed(MyplaceFeedDTO feedDto) throws AddException {
-        MyplaceFeed mf = MyplaceFeed
-                .builder()
-                .id(feedDto.getMyplaceId())
-//                .memberNickname(getMemberNickname(feedDto))
-                .content(feedDto.getContent())
-                .build();
-        mfr.save(mf);
-
+        MyplaceFeed mf = mapper.MyplaceFeedDTOtoMyplaceFeed(feedDto);
+        try {
+            mfr.save(mf);
+        } catch (DataAccessException e) {
+            throw new AddException("피드 생성 중 오류가 발생했습니다");
+        }
     }
 
 
@@ -71,16 +61,21 @@ public class MyplaceFeedService {
      * @throws AddException
      */
     public void updateMyFeed(MyplaceFeedDTO feedDto) throws FindException, ModifyException {
-        MyplaceFeedDTO mfDto = findMyFeedById(feedDto.getMyplaceId());
-
-        MyplaceFeed mf = MyplaceFeed
-                .builder()
-                .id(mfDto.getMyplaceId())
-//                .memberNickname(getMemberNickname(mfDto))
-                .content(feedDto.getContent())
-                .build();
-
-        mfr.save(mf);
+        Optional<MyplaceFeed> feed = mfr.findById(feedDto.getMyplaceId());
+        if (feed.isPresent()) {
+            MyplaceFeed updatedFeed = MyplaceFeed.builder()
+                    .myplaceId(feed.get().getMyplaceId())
+                    .memberEmail(feed.get().getMemberEmail())
+                    .content(feedDto.getContent())
+                    .build();
+            try {
+                mfr.save(updatedFeed);
+            } catch (DataAccessException e) {
+                throw new ModifyException("피드 수정 중 오류가 발생했습니다");
+            }
+        } else {
+            throw new FindException("해당 피드를 찾을 수 없습니다");
+        }
     }
 
     /**
@@ -88,8 +83,17 @@ public class MyplaceFeedService {
      * @param feedId 마이플레이스id
      * @throws RemoveException
      */
-    public void deleteMyFeed(Long feedId) throws RemoveException {
-        mfr.deleteById(feedId);
+    public void deleteMyFeed(Long feedId) throws RemoveException, FindException {
+        Optional<MyplaceFeed> feed = mfr.findById(feedId);
+        if (feed.isPresent()) {
+            try{
+                mfr.deleteById(feedId);
+            } catch(EmptyResultDataAccessException e){
+                throw new RemoveException("피드 삭제 중 오류가 발생했습니다");
+            }
+        } else {
+            throw new FindException("해당 피드를 찾을 수 없습니다");
+        }
     }
 
 }
