+++ src/main/java/com/kosa/mapbegood/domain/ourmap/ourplaceFeed/service/OurplaceFeedService.java
@@ -1,25 +1,30 @@
 package com.kosa.mapbegood.domain.ourmap.ourplaceFeed.service;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import com.kosa.mapbegood.domain.ourmap.ourplaceFeed.dto.OurplaceFeedDTO;
 import com.kosa.mapbegood.domain.ourmap.ourplaceFeed.entity.OurplaceFeed;
 import com.kosa.mapbegood.domain.ourmap.ourplaceFeed.entity.OurplaceFeedEmbedded;
+import com.kosa.mapbegood.domain.ourmap.ourplaceFeed.mapper.OurplaceFeedMapper;
 import com.kosa.mapbegood.domain.ourmap.ourplaceFeed.repository.OurplaceFeedRepository;
 import com.kosa.mapbegood.exception.AddException;
 import com.kosa.mapbegood.exception.FindException;
 import com.kosa.mapbegood.exception.ModifyException;
 import com.kosa.mapbegood.exception.RemoveException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
 
 @Service
-public class OurplaceFeedService {
+public class OurplaceFeedService implements OurplaceFeedServiceInterface {
     @Autowired
     private OurplaceFeedRepository ofr;
+
+    @Autowired
+    private OurplaceFeedMapper mapper;
 
     /**
      * DTO 객체 타입의 그룹피드id(아워플레이스id, 닉네임)를 entity 타입으로 변환한다
@@ -27,10 +32,7 @@ public class OurplaceFeedService {
      * @return 복합키 엔티티
      */
     public OurplaceFeedEmbedded getOurFeedEmId(OurplaceFeedDTO feedDto) {
-        OurplaceFeedEmbedded ofid = new OurplaceFeedEmbedded();
-        ofid.setOurplaceId(feedDto.getOurplaceId());
-//        ofid.setNickname(feedDto.getMemberNickname());
-        return ofid;
+        return mapper.ourplaceFeedDTOtoOurplaceFeedEmbedded(feedDto);
     }
 
     /**
@@ -42,7 +44,7 @@ public class OurplaceFeedService {
     public List<OurplaceFeedDTO> findAllOurFeed(Iterable<OurplaceFeedEmbedded> ourplaceId) throws FindException {
         List<OurplaceFeed> ofList = ofr.findAllById(ourplaceId);
         List<OurplaceFeedDTO> ourFeedList = new ArrayList<>();
-        for(OurplaceFeed of: ofList){
+        for (OurplaceFeed of : ofList) {
             OurplaceFeedDTO ourFeedDto = findOurFeedById(of.getId());
             ourFeedList.add(ourFeedDto);
         }
@@ -57,12 +59,11 @@ public class OurplaceFeedService {
      */
     public OurplaceFeedDTO findOurFeedById(OurplaceFeedEmbedded feedId) throws FindException {
         Optional<OurplaceFeed> feed = ofr.findById(feedId);
-        OurplaceFeedDTO feedDto = new OurplaceFeedDTO();
-        feedDto.setOurplaceId(feedId.getOurplaceId());
-        feedDto.setMemberEmail(feedId.getEmail());
-        feedDto.setContent(feed.get().getContent());
-        feedDto.setCreatedAt(feed.get().getCreatedAt());
-        return feedDto;
+        if (feed.isPresent()) {
+            return mapper.OurplaceFeedToOurplaceFeedDTO(feed.get());
+        } else {
+            throw new FindException("해당 피드를 찾을 수 없습니다");
+        }
     }
 
     /**
@@ -71,13 +72,12 @@ public class OurplaceFeedService {
      * @throws AddException
      */
     public void createOurFeed(OurplaceFeedDTO feedDto) throws AddException {
-        OurplaceFeed of = OurplaceFeed
-                .builder()
-                .id(getOurFeedEmId(feedDto))
-                .content(feedDto.getContent())
-                .build();
-        ofr.save(of);
-
+        OurplaceFeed of = mapper.OurplaceFeedDTOtoOurplaceFeed(feedDto);
+        try {
+            ofr.save(of);
+        } catch (DataAccessException e) {
+            throw new AddException("피드 생성 중 오류가 발생했습니다");
+        }
     }
 
     /**
@@ -87,15 +87,21 @@ public class OurplaceFeedService {
      * @throws AddException
      */
     public void updateOurFeed(OurplaceFeedDTO feedDto) throws FindException, ModifyException {
-        OurplaceFeedDTO ofDto = findOurFeedById(getOurFeedEmId(feedDto));
-
-        OurplaceFeed of = OurplaceFeed
-                .builder()
-                .id(getOurFeedEmId(ofDto))
-                .content(feedDto.getContent())
-                .build();
-
-        ofr.save(of);
+        Optional<OurplaceFeed> feed = ofr.findById(getOurFeedEmId(feedDto));
+        if(feed.isPresent()) {
+            OurplaceFeed updatedFeed = OurplaceFeed
+                    .builder()
+                    .id(feed.get().getId())
+                    .content(feedDto.getContent())
+                    .build();
+            try {
+                ofr.save(updatedFeed);
+            } catch (DataAccessException e) {
+                throw new ModifyException("피드 수정 중 오류가 발생했습니다");
+            }
+        }else{
+            throw new FindException("해당 피드를 찾을 수 없습니다");
+        }
     }
 
     /**
@@ -103,8 +109,17 @@ public class OurplaceFeedService {
      * @param feedId 피드id(아워플레이스id, 닉네임)
      * @throws RemoveException
      */
-    public void deleteOurFeed(OurplaceFeedEmbedded feedId) throws RemoveException {
-        ofr.deleteById(feedId);
+    public void deleteOurFeed(OurplaceFeedEmbedded feedId) throws RemoveException, FindException {
+        Optional<OurplaceFeed> feed = ofr.findById(feedId);
+        if (feed.isPresent()) {
+            try{
+                ofr.deleteById(feedId);
+            } catch(EmptyResultDataAccessException e){
+                throw new RemoveException("피드 삭제 중 오류가 발생했습니다");
+            }
+        } else {
+            throw new FindException("해당 피드를 찾을 수 없습니다");
+        }
     }
 
 }
